export the HTML block
The export() method in blocks has been renamed to exportProperties()
This makes it clear that it only exposes the metadata of a block. A
new exportContents() method does now exist that exposes the contents
of a block. If the so exposed data is not null, it will be added as
a CDATA section to the block node's content.