fix navigation when plugin isn't the overview
Due to misconfigurations in the Stud.IP course categories, a course
using the Mooc.IP plugin might not be responsible to provide the course
overview tab. In this case, the course view must not terminate with an
exception since then the user wouldn't be able to activate the overview
again. Instead, the plugin will fall back to a some basic navigation.